Refacotoring container code
