+++ terra/tests/test_compute_singularity.py
@@ -1,0 +1,20 @@
+# import os
+# import re
+# import json
+from unittest import mock
+# import warnings
+
+# from terra import settings
+# from terra.compute import base
+# from terra.compute import docker
+# from terra.compute import compute
+# import terra.compute.utils
+
+from .utils import TestCase
+
+
+class TestComputeSingularCase(TestCase):
+  pass
+
+class TestSingular(TestComputeSingularCase):
+  pass
